refactor: 사이드바 컴포넌트별 분리
